requirements.txt
Django==3.2.13
djangorestframework==3.13.1
django-filter==21.1
coverage==6.4.1
requests==2.28.1
celery==5.1.2
redis==3.5.3
coverage==6.4.1